\name{PIPS-internal}
%% User-level objects not exported to the namespace:
\alias{t.test.mult}
\alias{prop.test.mult}
\alias{sim.basic}

\title{Internal PIPS Functions}
\description{
  Internal PIPS functions
}
\details{
  User-level objects not exported and not meant to be called by the user.
}
\keyword{ internal }
